+++ apps/modernization-ui/src/apps/page-builder/page/management/Edit/header/PageHeader.module.scss
@@ -12,7 +12,7 @@
     border-radius: 0.3125rem;
     background: colors.$base-white;
 
-    margin: 1rem 0;
+    margin: 1.5rem 0;
     @extend %thin;
 
     .infoAndButtonWrapper {
