+++ apps/modernization-ui/src/apps/page-builder/page/management/header/tabs/Tabs.module.scss
@@ -1,0 +1,37 @@
+@use 'styles/colors';
+
+.tabs {
+    display: flex;
+    gap: 2.5rem;
+    list-style-type: none;
+    margin: 0;
+    padding: 0;
+
+    .tab {
+        cursor: pointer;
+
+        &:focus {
+            outline: 0.25rem solid #2491ff;
+        }
+
+        .tabLink {
+            display: block;
+            color: colors.$base-black;
+            font-size: 1rem;
+            font-weight: 400;
+            margin: 0;
+            padding: 0.6rem;
+
+            &.selected {
+                border-bottom: solid 0.4375rem colors.$primary;
+                font-weight: 700;
+            }
+
+            &:hover {
+                border-bottom: solid 0.4375rem colors.$primary;
+                text-decoration: none;
+                color: colors.$primary;
+            }
+        }
+    }
+}
